+++ psamm/gapfill.py
@@ -29,7 +29,7 @@ class GapFillError(Exception):
     """Indicates an error while running GapFind/GapFill"""
 
 
-def gapfind(model, solver, epsilon=1e-5, v_max=1000):
+def gapfind(model, solver, epsilon=0.001, v_max=1000):
     """Identify compounds in the model that cannot be produced.
 
     Yields all compounds that cannot be produced. This method
@@ -100,7 +100,7 @@ def gapfind(model, solver, epsilon=1e-5, v_max=1000):
             yield compound
 
 
-def gapfill(model, core, blocked, solver, epsilon=1e-5, v_max=1000):
+def gapfill(model, core, blocked, solver, epsilon=0.001, v_max=1000):
     """Find a set of reactions to add such that no compounds are blocked.
 
     Returns two iterators: first an iterator of reactions not in
@@ -147,20 +147,20 @@ def gapfill(model, core, blocked, solver, epsilon=1e-5, v_max=1000):
             prob.add_linear_constraints(v >= model.limits[reaction_id].lower)
         else:
             ym = prob.var(('ym', reaction_id))
-            prob.add_linear_constraints(v >= -v_max*ym)
+            prob.add_linear_constraints(v >= -v_max * ym)
         prob.add_linear_constraints(v <= model.limits[reaction_id].upper)
 
     # Add constraints on database reactions
     for reaction_id in database_reactions:
         v = prob.var(('v', reaction_id))
         yd = prob.var(('yd', reaction_id))
-        prob.add_linear_constraints(v >= yd * model.limits[reaction_id].lower)
-        prob.add_linear_constraints(v <= yd * model.limits[reaction_id].upper)
+        prob.add_linear_constraints(
+            v >= yd * model.limits[reaction_id].lower,
+            v <= yd * model.limits[reaction_id].upper)
 
     # Define constraints on production of blocked metabolites in reaction
     binary_cons_lhs = {compound: 0 for compound in blocked}
-    for spec, value in iteritems(model.matrix):
-        compound, reaction_id = spec
+    for (compound, reaction_id), value in iteritems(model.matrix):
         if compound in blocked and value != 0:
             prob.define(('w', reaction_id, compound),
                         types=lp.VariableType.Binary)
@@ -168,11 +168,21 @@ def gapfill(model, core, blocked, solver, epsilon=1e-5, v_max=1000):
             w = prob.var(('w', reaction_id, compound))
             sv = float(value) * prob.var(('v', reaction_id))
 
-            prob.add_linear_constraints(sv >= epsilon-v_max*(1 - w))
-            prob.add_linear_constraints(sv <= v_max*w)
+            prob.add_linear_constraints(sv >= epsilon - v_max * (1 - w))
+            prob.add_linear_constraints(sv <= v_max * w)
 
-            if reaction_id in model.reversible or value > 0:
+            if model.is_reversible(reaction_id) or value > 0:
                 binary_cons_lhs[compound] += w
+            elif reaction_id in core:
+                # In this case, we need to perform a logical AND on the w and
+                # ym variables. This is done by introducing another helper
+                # variable, yn.
+                prob.define(('yn', reaction_id, compound),
+                            types=lp.VariableType.Binary)
+                yn = prob.var(('yn', reaction_id, compound))
+                prob.add_linear_constraints(
+                    2 * yn <= w + prob.var(('ym', reaction_id)))
+                binary_cons_lhs[compound] += yn
 
     for compound, lhs in iteritems(binary_cons_lhs):
         if compound in blocked:
@@ -195,12 +205,12 @@ def gapfill(model, core, blocked, solver, epsilon=1e-5, v_max=1000):
 
     def added_iter():
         for reaction_id in database_reactions:
-            if result.get_value(('yd', reaction_id)) > 0:
+            if result.get_value(('yd', reaction_id)) > 0.5:
                 yield reaction_id
 
     def reversed_iter():
         for reaction_id in core:
-            if result.get_value(('ym', reaction_id)) > 0:
+            if result.get_value(('ym', reaction_id)) > 0.5:
                 yield reaction_id
 
     return added_iter(), reversed_iter()