+++ psamm/lpsolver/generic.py
@@ -126,7 +126,12 @@ class Solver(BaseSolver):
         solver = max(solvers, key=lambda s: priority.get(s['name'], 0))
         logger.debug('Using solver {}'.format(solver['name']))
 
+        self._properties = solver
         self._solver = solver['class']()
+
+    @property
+    def properties(self):
+        return self._properties
 
     def create_problem(self):
         """Create a :class:`Problem <psamm.lpsolver.lp.Problem>` instance"""
@@ -151,7 +156,7 @@ def parse_solver_setting(s):
     return key, value
 
 
-def list_solvers():
+def list_solvers(args=None):
     """Entry point for listing available solvers."""
     parser = argparse.ArgumentParser(
         description='''List LP solver available in PSAMM. This will produce a
@@ -165,10 +170,10 @@ def list_solvers():
     parser.add_argument(
         'requirement', nargs='*', type=str,
         help='Additional requirements on the selected solvers')
-    args = parser.parse_args()
+    parsed_args = parser.parse_args(args)
 
     requirements = {}
-    for arg in args.requirement:
+    for arg in parsed_args.requirement:
         try:
             key, value = parse_solver_setting(arg)
         except ValueError as e: