+++ psamm/lpsolver/qsoptex.py
@@ -135,12 +135,7 @@ class Problem(BaseProblem):
         constraints = []
 
         for relation in relations:
-            self._check_relation(relation)
-            if isinstance(relation, bool):
-                # A bool in place of a relation is accepted to mean
-                # a relation that does not involve any variables and
-                # has therefore been evaluated to a truth-value (e.g
-                # '0 == 0' or '2 >= 3').
+            if self._check_relation(relation):
                 constraints.append(Constraint(self, None))
             else:
                 for name in self._add_constraints(relation):
@@ -218,7 +213,14 @@ class Result(BaseResult):
     def success(self):
         """Return boolean indicating whether a solution was found"""
         self._check_valid()
-        return self._problem._p.get_status() == 1
+        return self._problem._p.get_status() == qsoptex.SolutionStatus.OPTIMAL
+
+    @property
+    def unbounded(self):
+        """Whether the solution is unbounded"""
+        self._check_valid()
+        return (self._problem._p.get_status() ==
+                qsoptex.SolutionStatus.UNBOUNDED)
 
     @property
     def status(self):