+++ psamm/tests/test_fluxanalysis.py
@@ -24,6 +24,8 @@ from psamm import fluxanalysis
 from psamm.datasource.modelseed import parse_reaction
 from psamm.lpsolver import generic
 
+from six import itervalues
+
 
 class TestFluxBalance(unittest.TestCase):
     def setUp(self):
@@ -123,6 +125,37 @@ class TestFluxBalance(unittest.TestCase):
         self.assertAlmostEqual(p.get_flux('rxn_5'), 1000)
         self.assertAlmostEqual(p.get_flux('rxn_6'), 1000)
 
+    def test_flux_balance_infeasible(self):
+        self.model.limits['rxn_6'].lower = 500
+        self.model.limits['rxn_1'].upper = 10
+        p = fluxanalysis.FluxBalanceProblem(self.model, self.solver)
+
+        try:
+            p.maximize('rxn_6')
+        except fluxanalysis.FluxBalanceError as e:
+            self.assertFalse(e.result.unbounded)
+        else:
+            self.fail('FluxBalanceError was not raised!')
+
+    def test_flux_balance_unbounded(self):
+        unbounded = float('-inf'), float('inf')
+        self.model.limits['rxn_1'].bounds = unbounded
+        self.model.limits['rxn_3'].bounds = unbounded
+        self.model.limits['rxn_6'].bounds = unbounded
+        p = fluxanalysis.FluxBalanceProblem(self.model, self.solver)
+
+        if self.solver.properties['name'] == 'qsoptex':
+            # QSopt_ex returns status code 100 for this example. It seems that
+            # it is unable to determine whether the problem is unbounded.
+            self.skipTest('Skipping because of known issue with QSopt_ex')
+
+        try:
+            p.maximize('rxn_6')
+        except fluxanalysis.FluxBalanceError as e:
+            self.assertTrue(e.result.unbounded)
+        else:
+            self.fail('FluxBalanceError was not raised!')
+
 
 class TestFluxBalanceThermodynamic(unittest.TestCase):
     def setUp(self):
@@ -181,6 +214,9 @@ class TestFluxVariability(unittest.TestCase):
             self.model, self.model.reactions, {'rxn_6': 200},
             tfba=False, solver=self.solver))
 
+        for bounds in itervalues(fluxes):
+            self.assertEqual(len(bounds), 2)
+
         self.assertAlmostEqual(fluxes['rxn_1'][0], 100)
 
         self.assertAlmostEqual(fluxes['rxn_2'][0], 0)
@@ -193,6 +229,30 @@ class TestFluxVariability(unittest.TestCase):
 
         self.assertGreater(fluxes['rxn_7'][1], 0)
         self.assertGreater(fluxes['rxn_8'][1], 0)
+
+    def test_flux_variability_unbounded(self):
+        self.model.limits['rxn_7'].upper = float('inf')
+        self.model.limits['rxn_8'].upper = float('inf')
+
+        if self.solver.properties['name'] == 'qsoptex':
+            # QSopt_ex returns status code 100 for this example. It seems that
+            # it is unable to determine whether the problem is unbounded.
+            self.skipTest('Skipping because of known issue with QSopt_ex')
+
+        fluxes = dict(fluxanalysis.flux_variability(
+            self.model, self.model.reactions, {'rxn_6': 200},
+            tfba=False, solver=self.solver))
+
+        for bounds in itervalues(fluxes):
+            self.assertEqual(len(bounds), 2)
+
+        self.assertAlmostEqual(fluxes['rxn_6'][0], 200)
+
+        self.assertAlmostEqual(fluxes['rxn_7'][0], 0)
+        self.assertEqual(fluxes['rxn_7'][1], float('inf'))
+
+        self.assertAlmostEqual(fluxes['rxn_8'][0], 0)
+        self.assertEqual(fluxes['rxn_8'][1], float('inf'))
 
 
 class TestFluxVariabilityThermodynamic(unittest.TestCase):